feat(ntp-stats): hook MV3 desktop Sawtunaa/Basarunaa → pref C++
Bascule les compteurs cumulatifs locaux `kStatsMusicSecondsTotal` et
`kStatsPersonsBlurredTotal` de `Uint64` vers `Integer` pour permettre l'accès
via `chrome.settingsPrivate` depuis les extensions MV3 (settingsPrivate ne
sait pas mapper Uint64Pref → JS Number). 2^31 ≈ 68 ans de musique continue,
largement suffisant.

- stats_client.cc : `RegisterUint64Pref` → `RegisterIntegerPref`
- service.cc : `IncrementUint64Pref` → `IncrementIntegerPref` avec saturation
  à INT_MAX au lieu de UINT64_MAX
- handler.cc : `GetUint64` → `GetInteger`
- android.cc : `ReadLocalStateUint64` → `ReadLocalStateInteger`, jlong cast
  conservé pour exposition Java (Integer ≤ jlong)
- brave_prefs_util.cc : allowlist `kStatsMusicSecondsTotal` /
  `kStatsPersonsBlurredTotal` en PrefType::kNumber pour
  `chrome.settingsPrivate`. PrefsUtil::FindServiceForPref résout local_state
  si pref absente du profile, donc nos prefs en local_state passent.
- extensions/BUILD.gn : ajout dep `//brave/components/browther_analytics`

Côté MV3 desktop (Sawtunaa + Basarunaa, voir browther-private) : background
SW intercepte les messages `increment-{music-seconds,persons-blurred}-total`
et fait un read-modify-write via `chrome.settingsPrivate`. Validé device Mac
2026-06-08 : musique YouTube et personnes flouées s'incrémentent en temps
réel sur la NTP.

+++ browser/ui/webui/brave_new_tab_page_refresh/new_tab_page_handler.cc
@@ -487,15 +487,16 @@ void NewTabPageHandler::GetShieldsStats(GetShieldsStatsCallback callback) {
   stats->bandwidth_saved_bytes = pref_service_->GetUint64(
       brave_perf_predictor::prefs::kBandwidthSavedBytes);
   // Browther stats — compteurs cumulatifs stockés dans local_state par
-  // BrowtherAnalyticsService (cf. components/browther_analytics).
+  // BrowtherAnalyticsService (cf. components/browther_analytics). Integer
+  // (Int32) plutôt que Uint64 pour partage avec chrome.settingsPrivate.
   auto* local_state = g_browser_process->local_state();
   stats->music_seconds_removed =
       local_state
-          ? local_state->GetUint64(browther_analytics::prefs::kStatsMusicSecondsTotal)
+          ? local_state->GetInteger(browther_analytics::prefs::kStatsMusicSecondsTotal)
           : 0;
   stats->persons_blurred =
       local_state
-          ? local_state->GetUint64(browther_analytics::prefs::kStatsPersonsBlurredTotal)
+          ? local_state->GetInteger(browther_analytics::prefs::kStatsPersonsBlurredTotal)
           : 0;
   std::move(callback).Run(std::move(stats));
 }
